20220320 wireguard documentation
Adds `-t` (allocate pseudo tty) flag to `docker exec` call to run
`tcpdump` inside the WireGuard container. This speeds up delivery of
packet-capture output to the terminal.

Adds a note to the test for checking the router's public port to
explain that routers often return misleading information as a
security feature, meaning the test will not be definitive.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>