2023-02-14 MariaDB Dockerfile - experimental branch - PR 3 of 3
Adjusts Dockerfile to cope with rename of internal file path
`/defaults/my.cnf` to `/defaults/custom.cnf`.

Strategy is to check for the presence of both files and then apply the
stability patches to the first match. This should mean that the
Dockerfile is backwards compatible with images that use `my.cnf`.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>