2023-10-24 mjpg-streamer container - old-menu branch - PR 2 of 2
Adds `mjpg-streamer` service definition. The container is designed to
work with MotionEye. It provides a way to connect both "ribbon" and
USB-connected web cams. From the perspective of MotionEye, all cameras
are "networked" and accessed via HTTP. Streamer containers can be
running on the same or different hosts as MotionEye. While the most
common home arrangement is likely to be a single MotionEye instance
taking feeds from multiple streamers, there can also be many-to-many
relationships between streamer containers and MotionEye containers.

Documentation added to master branch.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>